+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [2.6.1] - 2018-04-22
 
 * Force better (imho) layouts for WooCommerce.
+* Register and display shop sidebar widget area for WooCommerce.
 * Add front-page widget areas.
 * Add line awesome font shortcodes.
 * Add first, last, and widget-# CSS classes to widgets.